+++ tests/test_pytest_plugin.py
@@ -33,8 +33,6 @@ def test_plugin(testdir: Pytester) -> None:
 
     testdir.makepyfile(
         """
-        import asyncio
-
         import pytest
         import sniffio
         from hypothesis import strategies, given
@@ -83,7 +81,7 @@ def test_asyncio(testdir: Pytester) -> None:
             def callback():
                 raise RuntimeError('failing fixture setup')
 
-            asyncio.get_event_loop().call_soon(callback)
+            asyncio.get_running_loop().call_soon(callback)
             await asyncio.sleep(0)
             yield None
 
@@ -93,7 +91,7 @@ def test_asyncio(testdir: Pytester) -> None:
                 raise RuntimeError('failing fixture teardown')
 
             yield None
-            asyncio.get_event_loop().call_soon(callback)
+            asyncio.get_running_loop().call_soon(callback)
             await asyncio.sleep(0)
         """
     )
@@ -124,7 +122,7 @@ def test_asyncio(testdir: Pytester) -> None:
                 raise Exception('foo')
 
             started = False
-            asyncio.get_event_loop().call_soon(callback)
+            asyncio.get_running_loop().call_soon(callback)
             await asyncio.sleep(0)
             assert started
 
@@ -197,6 +195,53 @@ def test_cancel_scope_in_asyncgen_fixture(testdir: Pytester) -> None:
         @pytest.mark.anyio
         async def test_cancel_in_asyncgen_fixture(asyncgen_fixture):
             assert asyncgen_fixture == 1
+        """
+    )
+
+    result = testdir.runpytest_subprocess(*pytest_args)
+    result.assert_outcomes(passed=len(get_all_backends()))
+
+
+def test_module_scoped_task_group_fixture(testdir: Pytester) -> None:
+    testdir.makeconftest(
+        """
+        import pytest
+
+        from anyio import CancelScope, create_memory_object_stream, create_task_group
+
+
+        @pytest.fixture(scope="module")
+        async def task_group():
+            async with create_task_group() as tg:
+                yield tg
+
+
+        @pytest.fixture
+        async def streams(task_group):
+            async def echo_messages(*, task_status):
+                with CancelScope() as cancel_scope:
+                    task_status.started(cancel_scope)
+                    async for obj in receive1:
+                        await send2.send(obj)
+
+            send1, receive1 = create_memory_object_stream()
+            send2, receive2 = create_memory_object_stream()
+            cancel_scope = await task_group.start(echo_messages)
+            yield send1, receive2
+            cancel_scope.cancel()
+        """
+    )
+
+    testdir.makepyfile(
+        """
+        import pytest
+
+
+        @pytest.mark.anyio
+        async def test_task_group(streams):
+            send1, receive2 = streams
+            await send1.send("hello")
+            assert await receive2.receive() == "hello"
         """
     )